requirements.txt
asgiref==3.2.10
autopep8==1.5.3
colorama==0.4.3
Django==3.0.8
django-debug-toolbar==2.2
django-filter==2.3.0
django-modeladmin-reorder==0.3.1
https://github.com/darklow/django-suit/tarball/v2
numpy==1.19.1
pep8==1.7.1
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
qrcode==6.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
static3==0.7.0
webencodings==0.5.1
whitenoise==5.1.0
zipp==3.1.0
packaging==20.4
Pillow==7.2.0
dj-database-url==0.5.0
django-heroku
gunicorn==20.0.4
